% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUPA_psycho.R
\docType{data}
\name{AUPA_psycho}
\alias{AUPA_psycho}
\title{Psychological eating behavior data set}
\format{A data frame with 559 observations, (row names from 1 to 559) and 31 items. The name of the items refers to the corresponding block (E, X, R, P, S).}
\usage{
data(AUPA_psycho)
}
\description{
The psychological behaviour items in this dataset is a part of French Research Project (AUPALESENS, 2010-2013) dealing with food behaviour and nutritional status of elderly people. 
There are 31 psychological items organised into five blocks, each aiming to describe a given behavioural characteristic: 
emotional eating (E) with six items, 
external eating (X) with five items, 
restricted eating (R) with five items, 
pleasure for food (P) with five items, 
and self esteem (S) with ten items. 
Detailed description and analysis of the emotional, external and restricted eating items for this study are available in Bailly, Maitre, Amand, Herve, and Alaphilippe (2012).
559 subjects were considered.
}
\examples{
   X = data(AUPA_psycho)
}
\references{
Bailly N, Maitre I, Amand M, Herve C, Alaphilippe D (2012). The Dutch Eating Behaviour Questionnaire(DEBQ). Assessment of eating behaviour in an aging French population. Appetite, 59(853-858).
}
\keyword{datasets}
