% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.scale.R
\name{block.scale}
\alias{block.scale}
\title{Scaling of a three-way array}
\usage{
block.scale(X, xcenter = TRUE, xscale = 0)
}
\arguments{
\item{X}{: a three-way array}

\item{xcenter}{: centering of X. By default X will be centered for both mode 2 and 3 (xcenter=TRUE), otherwise xcenter=FALSE}

\item{xscale}{: scaling parameter applied to X. By default no scaling (xscale=0) \cr
0 : no scaling only centering - the default \cr
1 : scaling with standard deviation of  (mode 2 x mode 3) elements \cr
2 : global scaling (each block i.e. each mode 2 slice will have the same inertia ) \cr
3 : global scaling (each block i.e. each mode 3 slice will have the same inertia )}
}
\value{
Xscaled     : the scaled three-way array
}
\description{
Centering and scaling of a three-way array
}
