% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biplot.R
\name{data_biplot}
\alias{data_biplot}
\title{biplot for the dataset}
\usage{
data_biplot(X, sX = TRUE, axeh = 1, axev = 2, cex.lab = 1)
}
\arguments{
\item{X}{the data matrix}

\item{sX}{TRUE/FALSE : standardization or not of the columns X (TRUE by default)}

\item{axeh}{component number for the horizontal axis}

\item{axev}{component number for the vertical axis}

\item{cex.lab}{: magnification to be used for labels (1 by default)}
}
\description{
Loading plot of the variables from a Principal Components Analysis.
scores of the observations are surimposed
}
