% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_comp.R
\name{get_comp}
\alias{get_comp}
\title{To get the latent variables associated with each cluster}
\usage{
get_comp(resclv, K = NULL)
}
\arguments{
\item{resclv}{: result of CLV(), CLV_kmeans() or LCLV()}

\item{K}{: the number of groups chosen (already defined if CLV_kmeans is used)}
}
\value{
\item{comp}{the group latent variables (centered, but not standardized) \cr
     For results of LCLV, two types of latent variables are available : \cr
     compt : The latent variables of the clusters defined according to the Xr variables, \cr
     compc : The latent variables of the clusters defined according to the Xu variables
     }
}
\description{
To get the latent variables associated with each cluster
}
\examples{
data(apples_sh)
resclvX <- CLV(X = apples_sh$senso, method = "directional", sX = TRUE)
comp4G<-get_comp(resclvX, K = 4)
}

