% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loading.R
\name{get_loading}
\alias{get_loading}
\title{Loadings of the variables on the latent component, in each cluster.}
\usage{
get_loading(resclv, K = NULL, type = "list", graph = FALSE,
  cex.lab = 1)
}
\arguments{
\item{resclv}{: result of CLV(), CLV_kmeans(), LCLV(), CLV3W() or CLV3W_kmeans()}

\item{K}{: the number of clusters chosen (already defined if CLV_kmeans or CLV3W_kmeans is used)}

\item{type}{: outputs in the form of a "list" (one element by cluster, by default) or a "vector" (available only if "sparselv" strategy is used)}

\item{graph}{: boolean, if TRUE, the barplot associated with the scores is displayed (default : graph=FALSE)}

\item{cex.lab}{: magnification to be used for labels (1 by default)}
}
\value{
\item{loading}{the loadings of the variables on each cluster's latent component}
}
\description{
To get the variables loadings for the latent component in each cluster. \cr
For CLV(_kmeans), the loadings are of particular interest when method="directional" or when strategy="sparse LV" \cr
For CLV3W(_kmeans), the loadings are given for the variables associated with mode 2 of the 3-way array.
}
