% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition.R
\name{get_partition}
\alias{get_partition}
\title{Clusters of variables.}
\usage{
get_partition(resclv, K = NULL, type = "vector")
}
\arguments{
\item{resclv}{: result of CLV(), CLV_kmeans() or LCLV()}

\item{K}{: the number of groups chosen (already defined if CLV_kmeans is used)}

\item{type}{: presented in the form of a "vector" (by default) or a "matrix"}
}
\value{
\item{partition}{the group's membership for the variables) }
}
\description{
To get the clusters of variables. This function returns the group's membership for the p variables. 
The output can be a vector p x 1 of integers between 1 and K, 
or a binary matrix of size p x n.
}
\examples{
data(apples_sh)
resclvX <- CLV(X = apples_sh$senso, method = "directional", sX = TRUE)
parti4G<-get_partition(resclvX, K = 4) 

}
