% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imput_clv.R
\name{imput_clv}
\alias{imput_clv}
\title{Imputation of a data matrix based on CLV results}
\usage{
imput_clv(x, X0, K = NULL)
}
\arguments{
\item{x}{: an object of class \code{clv}}

\item{X0}{: the initial data matrix with missing values (NA)}

\item{K}{: the number of Latent Variables to be considered, each of them being associated with a group of variables.}
}
\value{
X0imput : the imputed data matrix, in the original scale

Ximput  : the imputed matrix, centered and scaled according to the pretratment parameters chosen in CLV
}
\description{
For each variable, its missing data will be imputed according to the values of the latent variable of the group in which the variable belong to.
}
\details{
It is adviced to use a larger number of latent variables, on the basis of which the imputation will be done, than the suspected 'true' number of groups of variables
}
