% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clv.R
\name{plot.clv}
\alias{plot.clv}
\title{Graphical representation of the CLV clustering stages}
\usage{
\method{plot}{clv}(x, type = "dendrogram", cex = 0.8, ...)
}
\arguments{
\item{x}{: an object of class \code{clv}}

\item{type}{: What to plot. \cr 
"dendrogram" : the dendrogram of the hierchical clustering algorithm, \cr    
"delta"      : a barplot showing the variation of the clustering criterium after consolidation.}

\item{cex}{: Character expansion for labels.}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function plots either the CLV dendrogram or the variations of the consolidated CLV criterion.
}
\seealso{
CLV
}
