% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clv3w.R
\name{plot.clv3w}
\alias{plot.clv3w}
\title{Graphical representation of the CLV3W hierarchical clustering stages}
\usage{
\method{plot}{clv3w}(x, type = "dendrogram", cex = 0.8, ...)
}
\arguments{
\item{x}{: an object of class \code{clv3w}}

\item{type}{: What to plot. \cr
"dendrogram" : the dendrogram of the hierchical clustering algorithm, \cr
"delta"      : a barplot showing the variation of the clustering criterium after consolidation.}

\item{cex}{: Character expansion for labels.}

\item{\dots}{Additional arguments passed on to the real \code{print}.}
}
\description{
This function plots either the CLV3W dendrogram or the variations of the consolidated CLV3W criterion.
}
\seealso{
CLV3W
}
