% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lmclv.R
\name{predict.lmclv}
\alias{predict.lmclv}
\title{prediction for lmCLV models.}
\usage{
\method{predict}{lmclv}(object, newdata, shrinkp, ...)
}
\arguments{
\item{object}{: result of class \code{lmclv}}

\item{newdata}{: Data frame of observations for which to make predictions}

\item{shrinkp}{: shrinkage  parameter.}

\item{...}{: further arguments passed to or from other methods}
}
\value{
a matrix of the predicted values, \cr
        each column with an increasing number of CLV component included \cr 
        the first column being for the null model \cr
        if the response if a binary factor, two additional matrices are provided :\cr
        the probabilities of belonging to class 1 and the response values (0 or 1).
}
\description{
To get the predicted response values based on lmCLV model.
}
