% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.clv.r
\name{print.clv}
\alias{print.clv}
\title{Print the CLV results}
\usage{
\method{print}{clv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{clv}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the CLV results
}
\seealso{
CLV
}
