\name{print.lclv}
\alias{print.lclv}
\title{Print the LCLV results}
\usage{
\method{print}{lclv}(x, ...)
}
\arguments{
  \item{x}{an object of class lclv}

  \item{...}{further arguments passed to or from other
  methods}
}
\description{
Print the LCLV results
}
\seealso{
LCLV
}

