% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_quali.R
\name{stand_quali}
\alias{stand_quali}
\title{Standardization of the qualitative variables}
\usage{
stand_quali(X.quali, metric = "chisq")
}
\arguments{
\item{X.quali}{: a factor or a data frame with several factors}

\item{metric}{: the metric to be used, i.e. each category is weighted by the inverse of the square-root of its relative frequency}
}
\value{
Xdisj.sd : a standardized matrix with as many columns as categories associated with the qualitative variables.
}
\description{
pretreatment of qualitative variables
}
