% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clv.R
\name{summary.clv}
\alias{summary.clv}
\title{summary and description of the clusters of variables}
\usage{
\method{summary}{clv}(object, K = NULL, ...)
}
\arguments{
\item{object}{: result of CLV() or CLV_kmeans()}

\item{K}{: the number of clusters (unless if CLV_kmeans was used)}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function provides the list of the variables within each group and complementary informations.
Users will be asked to specify the number of clusters,
}
\details{
The ouputs include :
\itemize{
\item the size of the groups, \cr
\item the list of the variables within each group. FFor each cluster, the correlation of the each variable with its group latent component 
  and the correlation with the next neighbouring group latent component are given.  \cr
\item the proportion of the variance within each group explained by its latent variable, \cr
 \item the proportion of the whole dataset account by the group latent variables \cr
 \item the matrix of correlation between the latent variables.}
}
