% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clv3w.R
\name{summary.clv3w}
\alias{summary.clv3w}
\title{Summary and description of the clusters of (mode 2) variables associated with CLV3W or CLV3W_kmeans}
\usage{
\method{summary}{clv3w}(object, K = NULL, ...)
}
\arguments{
\item{object}{: result of CLV3W() or CLV3W_kmeans()}

\item{K}{: the number of clusters (unless if CLV3W_kmeans was used)}

\item{\dots}{Additional arguments passed on to the real \code{summary}.}
}
\description{
This function provides the list of the variables within each group and complementary informations.
Users will be asked to specify the number of clusters,
}
\details{
The ouputs include :
\itemize{
\item the size of the groups, \cr
\item the proportion of the variance within each group explained by its latent variable, \cr
 \item the proportion of the whole dataset accounted by the group latent variables \cr
\item the latent components (mode 1) associated to the various groups, \cr
\item the weights (mode 3) associated to the various groups, \cr
\item the list of the variables within each group. For each cluster, the loading (mode 2) of the variable is given together with the correlation of the block component with its group latent component
  and the correlation with the next neighbouring group latent component are given.  \cr
 \item the matrix of correlation between the latent variables.}
}
