% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLV_kmeans.R
\name{CLV_kmeans}
\alias{CLV_kmeans}
\title{K-means algorithm for the clustering of variables}
\usage{
CLV_kmeans(X, Xu = NULL, Xr = NULL, method, sX = TRUE, sXr = FALSE,
  sXu = FALSE, clust, iter.max = 20, nstart = 100,
  strategy = "none", rho = 0.3)
}
\arguments{
\item{X}{The matrix of the variables to be clustered}

\item{Xu}{The external variables associated with the columns of X}

\item{Xr}{The external variables associated with the rows of X}

\item{method}{The criterion to use in the cluster analysis.\cr 
1 or "directional" : the squared covariance is used as a measure of proximity (directional groups). \cr    
2 or "local"       : the covariance is used as a measure of proximity (local groups)}

\item{sX}{TRUE/FALSE : standardization or not of the columns X (TRUE by default)\cr
(predefined -> cX = TRUE : column-centering of X)}

\item{sXr}{TRUE/FALSE : standardization or not of the columns Xr (FALSE by default)\cr
(predefined -> cXr    = TRUE : column-centering of Xr)}

\item{sXu}{TRUE/FALSE : standardization or not of the columns Xu (FALSE by default)\cr
(predefined -> cXu= FALSE : no centering, Xu considered as a weight matrix)}

\item{clust}{: a number i.e.  the size of the partition, K,
or  a vector of INTEGERS i.e. the group membership of each variable in the initial partition (integer between 1 and K)}

\item{iter.max}{maximal number of iteration for the consolidation (20 by default)}

\item{nstart}{nb of random initialisations in the case where init is a number  (100 by default)}

\item{strategy}{"none" (by default), or "kplusone" (an additional cluster for the noise variables),
or "sparselv" (zero loadings for the noise variables)}

\item{rho}{a threshold of correlation between 0 and 1 (0.3 by default)}
}
\value{
\item{tabres}{ 
        The value of the clustering criterion at convergence.\cr
        The percentage of the explained initial criterion value.\cr
        The number of iterations in the partitioning algorithm.}
        \item{clusters}{ the group's membership}
        \item{comp}{ The latent components of the clusters}
        \item{loading}{ if there are external variables Xr or Xu :  The loadings of the external variables}
}
\description{
K-means algorithm for the clustering of variables. Directional or local groups may be defined. 
Each group of variables is associated with a latent component. 
Moreover external information collected on the observations or on the variables may be introduced.
}
\details{
The initalization can be made at random, repetitively, or can be defined by the user.

The parameter "strategy" makes it possible to choose a strategy for setting aside variables
that do not fit into the pattern of any cluster.
}
\examples{
data(apples_sh)
#local groups with external variables Xr 
resclvkmYX <- CLV_kmeans(X = apples_sh$pref, Xr = apples_sh$senso,method = "local",
          sX = FALSE, sXr = TRUE, clust = 2, nstart = 20)
}
\references{
Vigneau E., Qannari E.M. (2003). Clustering of variables around latents components. Comm. Stat, 32(4), 1131-1150.

Vigneau E., Chen M., Qannari E.M. (2015). ClustVarLV:  An R Package for the clustering of Variables around Latent Variables. The R Journal, 7(2), 134-148

Vigneau E., Chen M. (2016). Dimensionality reduction by clustering of variables while setting aside atypical variables. Electronic Journal of Applied Statistical Analysis, 9(1), 134-153
}
\seealso{
CLV, LCLV
}
