% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition.R
\name{get_partition}
\alias{get_partition}
\title{clusters memberships for a partition into K clusters.}
\usage{
get_partition(resclv, K = NULL, type = "vector")
}
\arguments{
\item{resclv}{: result of CLV(), CLV_kmeans(), LCLV(), CLV3W() or CLV3W_km()}

\item{K}{: the number of clusters chosen (already defined if CLV_kmeans or CLV3W_kmeans is used)}

\item{type}{: "vector" (by default) for output given as a vector of integers between 1 and K (with 0 for "kplusone" strategy), \cr
"matrix", the output given as a binary matrix of size p x n.}
}
\value{
\item{partition}{the group's membership for the variables, in a vector or matrix form. \cr
                     For CLV3W object, a vector of memberships with mode 2) }
}
\description{
To get the clusters memberships of the variables.
}
\examples{
data(apples_sh)
resclvX <- CLV(X = apples_sh$senso, method = "directional", sX = TRUE)
parti4G<-get_partition(resclvX, K = 4) 

}
