\name{print.clvkmeans}
\alias{print.clvkmeans}
\title{Print the CLV_kmeans results}
\usage{
\method{print}{clvkmeans}(x, ...)
}
\arguments{
  \item{x}{an object of class clvkmeans}

  \item{...}{further arguments passed to or from other
  methods}
}
\description{
Print the CLV_kmeans results
}
\seealso{
CLV_kmeans
}

