% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lclv.r
\name{print.lclv}
\alias{print.lclv}
\title{Print the LCLV results}
\usage{
\method{print}{lclv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{lclv}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the LCLV results
}
\seealso{
LCLV
}
